Add implementation for the Generic Merge Operator
This merge operator can be used when the merge operands are not the same
as the Put values and it is not possible to overwrite previous operands.

Current implementation provides a new merge-method, which takes a string
representing the merge "operation". Clients can define own logic to interpret
these operation in order to perform complex merging on existing data.

TODO: Investigate why the partial-merge function is never called.

NOTICE: Current implementation is not optimal, since calling "mergeObject:forKey:"
would cause a crash when:
- The generic merge operator is used AND
- The merged object is not a NSString
Possible fix: check for passed objects and ignore them accordingly.