Refactor options class into database, column family and encoding options classes
The RocksDBOptions class is just a composition of the db and column family
options, this somewhat reflects the underlying implementation.
The encoding options object is needed to provide the convenience conversion
blocks and it is wrapped by the column family options, since separate column
families can have different kinds of keys/values.