Add column-family-based methods to write batch class
A standalone write batch can only be created via a DB instance, a column
family is a DB instance as well. All previous CRUD methods operate
on the column family that was used to create the standalone write batch
object. New API allows for using the same write batch to operate on other
column families.